requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
textblob>=0.17.1
pandas>=2.1.0
rich>=13.7.0
httpx>=0.25.0
feedparser>=6.0.10
